Make DefaultRxRestClient log like DefaultRestClientRequest
In case we have really hard exceptions like

java.nio.channels.UnresolvedAddressException
	at sun.nio.ch.Net.checkAddress(Net.java:101)

which do not really give any information about the nature of the unresolved
host we need the information that is currently only available at the level of
the DefaultRxRestClient. So we at least log it (DefaultRestClientRequest does
so as well) that we know what was unresolvable.